stage2_feed_replay/requirements.txt
websockets>=12.0
requests>=2.32.0
pyserial>=3.5
